test-gen/pyproject.toml
[tool.poetry]
name = "test-gen"
version = "0.1.0"
description = ""
authors = ["Daniel Clark <daniel.clark@ibm.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
kr8s = "^0.13.3"
grpcio = "^1.60.1"
grpcio-tools = "^1.60.1"
progress = "^1.6"
"ruamel.yaml" = "^0.18.6"
urllib3 = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
